store/todo: tasks.pos is INTEGER in the substrate
Task.Pos was int, which lift's type table maps to TEXT; the ordering
spine (extra.sql's (pos, created_seq) index) would sort lexically the
day anything orders in SQL. int64 in the metadata, regenerated: ddl
emits INTEGER, the domain carries int64. Ordering today happens in Go,
so nothing regressed; the drift guard confirms byte-identity.

+++ store/todo/metadata/metadata.go
@@ -35,7 +35,7 @@ type Task struct {
 	ID           string `primary:"true" alias:"name=id,nullable=false"`
 	Text         string `alias:"name=text,nullable=false"`
 	Status       string `alias:"name=status,nullable=false"`
-	Pos          int    `alias:"name=pos,nullable=false"`
+	Pos          int64  `alias:"name=pos,nullable=false"`
 	CreatedSeq   int64  `alias:"name=created_seq,nullable=false"`
 	CreatedEvent Event  `link:"from=Event,on=seq,many=false"`
 	UpdatedSeq   int64  `alias:"name=updated_seq,nullable=false"`
